// x-pack/platform/plugins/shared/fleet/common/types/models/epm.d.ts
export declare enum KibanaAssetType {
    dashboard = "dashboard",
    lens = "lens",
    visualization = "visualization",
    search = "search",
    indexPattern = "index_pattern",
    map = "map",
    mlModule = "ml_module",
    securityAIPrompt = "security_ai_prompt",
    securityRule = "security_rule",
    cloudSecurityPostureRuleTemplate = "csp_rule_template",
    alertingRuleTemplate = "alerting_rule_template",
    sloTemplate = "slo_template",
    osqueryPackAsset = "osquery_pack_asset",
    osquerySavedQuery = "osquery_saved_query",
    tag = "tag"
}